+++ dubbo-registry/dubbo-registry-api/src/main/java/org/apache/dubbo/registry/client/DefaultServiceInstance.java
@@ -56,7 +56,7 @@ public class DefaultServiceInstance implements ServiceInstance {
     }
 
     public DefaultServiceInstance(String id, String serviceName, String host, Integer port) {
-        if (port != null && port.intValue() < 1) {
+        if (port != null && port < 1) {
             throw new IllegalArgumentException("The port must be greater than zero!");
         }
         this.id = id;
